requirements.txt
accelerate==0.21.0
bitsandbytes==0.41.0
cerebrium==1.3.1
pydantic==2.1.1
pillow==10.0.0
scipy==1.11.1
torch==2.0.1
transformers==4.31.0